Add softcore cutoff to Coulomb force.
This allows us to set a limit to the Coulomb force which can be necessary to
make naive integrators of the particle motion stable. Without such a cutoff the
Coulomb force can become arbitrarily large leading to accuracy and stability
issues in most integrators.

We bound the softcore radius from below by an amount that is designed to work
well with finite floating point precision. This allows us to get rid of the
check for identical particles in the force computation loop.